Update tools.sh - tool_update_cc_cfg add ability to make comment from existing to defaults
update tool_update_cc_cfg function with ability to make comment from existing values to defaults.

prevent tool_update_cc_cfg function to not update variable twice.

above changes are related to idea if primary chain configuration file is forked and secondary is using conf variable from primary, so we are able to change only few variables we need.

+++ src/tools.sh
@@ -460,15 +460,18 @@ function tool_update_cc_cfg() { # cc_main_cfg_path # cc_main_cfg_add # cc_ticker
     local naming_suffix=${3}
     local cc_action_cc_config=${4}
     
-        ## main wallet conf file load
-        echo "INFO >> ${naming_suffix} >> checking main cc config >> ${cc_main_cfg_path}"
-        (test ! -f "${cc_main_cfg_path}") && touch ${cc_main_cfg_path}
-        (test ! -f "${cc_main_cfg_path}") && echo "ERROR >> ${naming_suffix} >> ${cc_main_cfg_path} >> not exist" && exit 1
-        cc_main_cfg_data=$(<${cc_main_cfg_path})
-        
+    ## main wallet conf file load
+    echo "INFO >> ${naming_suffix} >> checking main cc config >> ${cc_main_cfg_path}"
+    (test ! -f "${cc_main_cfg_path}") && touch ${cc_main_cfg_path}
+    (test ! -f "${cc_main_cfg_path}") && echo "ERROR >> ${naming_suffix} >> ${cc_main_cfg_path} >> not exist" && exit 1
+    cc_main_cfg_data=$(<${cc_main_cfg_path})
+    
     # filter out comments in cc_main_cfg_add variable
     cc_main_cfg_add=`echo "${cc_main_cfg_add}" | grep -v "^#"`
     (test $? != 0) && echo "ERROR >> filter out comments in cc_main_cfg_add variable >> failed" && exit 1
+    
+    # test duplicates variable
+    cc_main_cfg_add_dup_test=""
     
     # go all config lines to be added
     for cfg_line in ${cc_main_cfg_add}; do
@@ -480,26 +483,59 @@ function tool_update_cc_cfg() { # cc_main_cfg_path # cc_main_cfg_add # cc_ticker
         
         # extract variable name and check if not empty
         var_name=`echo $cfg_line | cut -d "=" -f1`
+        var_value=`echo $cfg_line | cut -d "=" -f2`
         
-        if [ "${var_name}" != "" ]; then
+        # check variable to add duplicates, only first one value is applied
+        echo "${cc_main_cfg_add_dup_test}" | grep "=${var_name}=" > /dev/null && already_exist="1" || already_exist="0"
+        
+        if [[ "${var_name}" != "" ]] && [[ "${already_exist}" != "1"  ]]; then
             
             #~ echo "updating/adding variable >> $var_name"
             
+            # configuration update option
             if [ "${cc_action_cc_config}" = "update" ]; then
-                # if existing variable, comment it
-                # if commented variable, add new next to
-                cc_main_cfg_data=$(
-                echo "$cc_main_cfg_data" | sed \
-                    -e "s/^${var_name}=/#~ ${var_name}=/" \
-                    -e "0,/^#~ ${var_name}=/s/^#~ ${var_name}=.*/${cfg_line}\n&/"
-                )
-                (test $? != 0) && echo "ERROR >> update line >> ${cfg_line} >> ${cc_main_cfg_path} >> failed" && exit 1
+            
+                # configuration value is not empty so will be replaced
+                if [ "${var_value}" != "" ]; then
+                    # if existing variable, comment it
+                    # if commented variable, then add new variable next to
+                    cc_main_cfg_data=$(
+                    echo "$cc_main_cfg_data" | sed \
+                       -e "s/^${var_name}=/#~ ${var_name}=/" \
+                       -e "0,/^#~ ${var_name}=/s/^#~ ${var_name}=.*/${cfg_line}\n&/"
+                    )
+                    (test $? != 0) && echo "ERROR >> update line >> ${cfg_line} >> ${cc_main_cfg_path} >> failed" && exit 1
+               
+                # else configuration value is empty so will be just commented out
+                else
+                    # if existing variable, comment it
+                    # if commented variable, then add new commented out variable next to
+                    cc_main_cfg_data=$(
+                    echo "$cc_main_cfg_data" | sed \
+                       -e "s/^${var_name}=/#~ ${var_name}=/" \
+                       -e "0,/^#~ ${var_name}=/s/^#~ ${var_name}=.*/#~ ${cfg_line}\n&/"
+                    )
+                    (test $? != 0) && echo "ERROR >> update line >> ${cfg_line} >> ${cc_main_cfg_path} >> failed" && exit 1
+                fi
             fi
             
-            # if variable not added next to comment, so add new line(PLEASE LET NEWLINE THERE)
-            echo "${cc_main_cfg_data}" | grep "^${var_name}=" > /dev/null || cc_main_cfg_data="${cc_main_cfg_data}
+            # if variable in configuration still does not exist, then add it
+            if [ "${var_value}" != "" ]; then
+                # if variable not added next to comment, so add new line(PLEASE LET NEWLINE THERE)
+                echo "${cc_main_cfg_data}" | grep "^${var_name}=" > /dev/null || cc_main_cfg_data="${cc_main_cfg_data}
 ${cfg_line}"
-            (test $? != 0) && echo "ERROR >> add line >> ${cfg_line} >> ${cc_main_cfg_path} >> failed" && exit 1
+                (test $? != 0) && echo "ERROR >> add line >> ${cfg_line} >> ${cc_main_cfg_path} >> failed" && exit 1
+            
+            # if variable or commented variable still does not exist, then add variable as commented variable
+            else
+                # if variable not added next to comment, so add new line(PLEASE LET NEWLINE THERE)
+                echo "${cc_main_cfg_data}" | grep -e "^#~ ${var_name}=" -e "^${var_name}=" > /dev/null || cc_main_cfg_data="${cc_main_cfg_data}
+#~ ${cfg_line}"
+                (test $? != 0) && echo "ERROR >> add line >> ${cfg_line} >> ${cc_main_cfg_path} >> failed" && exit 1
+            fi
+            
+            # add into duplicates list check and prevent double update
+            cc_main_cfg_add_dup_test="${cc_main_cfg_add_dup_test}=${var_name}="
         fi
     done
     
@@ -529,4 +565,3 @@ if [ "$cc_dexsetup_dir" != "dexsetup" ]; then
     echo "ERROR >> dexsetup scripts must be executed from <dexsetup> directory"
     exit 1
 fi
-
